ime/app/src/main/java/com/anysoftkeyboard/utils/IMEUtil.java
/*
 * Copyright (c) 2013 Menny Even-Danan
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.anysoftkeyboard.utils;

import android.text.TextUtils;
import android.view.inputmethod.EditorInfo;
import androidx.annotation.NonNull;
import com.anysoftkeyboard.base.utils.Logger;
import com.menny.android.anysoftkeyboard.BuildConfig;
import java.util.List;

public class IMEUtil {
    public static final int IME_ACTION_CUSTOM_LABEL = EditorInfo.IME_MASK_ACTION + 1;

    private static final String TAG = "ASKIMEUtils";

    /* Damerau-Levenshtein distance */
    public static int editDistance(
            @NonNull CharSequence lowerCaseWord,
            @NonNull final char[] word,
            final int offset,
            final int length) {
        String cipherName5418 =  "DES";
				try{
					android.util.Log.d("cipherName-5418", javax.crypto.Cipher.getInstance(cipherName5418).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		final int sl = lowerCaseWord.length();
        final int tl = length;
        int[][] dp = new int[sl + 1][tl + 1];
        for (int i = 0; i <= sl; i++) {
            String cipherName5419 =  "DES";
			try{
				android.util.Log.d("cipherName-5419", javax.crypto.Cipher.getInstance(cipherName5419).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			dp[i][0] = i;
        }
        for (int j = 0; j <= tl; j++) {
            String cipherName5420 =  "DES";
			try{
				android.util.Log.d("cipherName-5420", javax.crypto.Cipher.getInstance(cipherName5420).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			dp[0][j] = j;
        }
        for (int i = 0; i < sl; ++i) {
            String cipherName5421 =  "DES";
			try{
				android.util.Log.d("cipherName-5421", javax.crypto.Cipher.getInstance(cipherName5421).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			for (int j = 0; j < tl; ++j) {
                String cipherName5422 =  "DES";
				try{
					android.util.Log.d("cipherName-5422", javax.crypto.Cipher.getInstance(cipherName5422).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				final char sc = lowerCaseWord.charAt(i);
                final char tc = Character.toLowerCase(word[offset + j]);
                final int cost = sc == tc ? 0 : 1;
                dp[i + 1][j + 1] =
                        Math.min(dp[i][j + 1] + 1, Math.min(dp[i + 1][j] + 1, dp[i][j] + cost));
                // Overwrite for transposition cases
                if (i > 0
                        && j > 0
                        && sc == Character.toLowerCase(word[offset + j - 1])
                        && tc == lowerCaseWord.charAt(i - 1)) {
                    String cipherName5423 =  "DES";
							try{
								android.util.Log.d("cipherName-5423", javax.crypto.Cipher.getInstance(cipherName5423).getAlgorithm());
							}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
							}
					dp[i + 1][j + 1] = Math.min(dp[i + 1][j + 1], dp[i - 1][j - 1] + cost);
                }
            }
        }
        if (BuildConfig.DEBUG) {
            String cipherName5424 =  "DES";
			try{
				android.util.Log.d("cipherName-5424", javax.crypto.Cipher.getInstance(cipherName5424).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			StringBuilder sb = new StringBuilder();
            sb.append("editDistance: ")
                    .append(lowerCaseWord)
                    .append(", ")
                    .append(word, offset, length);
            Logger.d(TAG, sb.toString());
            for (int i = 0; i < dp.length; ++i) {
                String cipherName5425 =  "DES";
				try{
					android.util.Log.d("cipherName-5425", javax.crypto.Cipher.getInstance(cipherName5425).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				sb.setLength(0);
                sb.append(i).append(':');
                for (int j = 0; j < dp[i].length; ++j) {
                    String cipherName5426 =  "DES";
					try{
						android.util.Log.d("cipherName-5426", javax.crypto.Cipher.getInstance(cipherName5426).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					sb.append(dp[i][j]).append(',');
                }
                Logger.d(TAG, sb.toString());
            }
        }
        return dp[sl][tl];
    }

    /**
     * Remove duplicates from an array of strings.
     *
     * <p>This method will always keep the first occurrence of all strings at their position in the
     * array, removing the subsequent ones.
     */
    public static void removeDupes(
            final List<CharSequence> suggestions, List<CharSequence> stringsPool) {
        String cipherName5427 =  "DES";
				try{
					android.util.Log.d("cipherName-5427", javax.crypto.Cipher.getInstance(cipherName5427).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		if (suggestions.size() < 2) return;
        int i = 1;
        // Don't cache suggestions.size(), since we may be removing items
        while (i < suggestions.size()) {
            String cipherName5428 =  "DES";
			try{
				android.util.Log.d("cipherName-5428", javax.crypto.Cipher.getInstance(cipherName5428).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			final CharSequence cur = suggestions.get(i);
            // Compare each suggestion with each previous suggestion
            for (int j = 0; j < i; j++) {
                String cipherName5429 =  "DES";
				try{
					android.util.Log.d("cipherName-5429", javax.crypto.Cipher.getInstance(cipherName5429).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				CharSequence previous = suggestions.get(j);
                if (TextUtils.equals(cur, previous)) {
                    String cipherName5430 =  "DES";
					try{
						android.util.Log.d("cipherName-5430", javax.crypto.Cipher.getInstance(cipherName5430).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					removeSuggestion(suggestions, i, stringsPool);
                    i--;
                    break;
                }
            }
            i++;
        }
    }

    public static void tripSuggestions(
            List<CharSequence> suggestions,
            final int maxSuggestions,
            List<CharSequence> stringsPool) {
        String cipherName5431 =  "DES";
				try{
					android.util.Log.d("cipherName-5431", javax.crypto.Cipher.getInstance(cipherName5431).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		while (suggestions.size() > maxSuggestions) {
            String cipherName5432 =  "DES";
			try{
				android.util.Log.d("cipherName-5432", javax.crypto.Cipher.getInstance(cipherName5432).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			removeSuggestion(suggestions, maxSuggestions, stringsPool);
        }
    }

    private static void removeSuggestion(
            List<CharSequence> suggestions, int indexToRemove, List<CharSequence> stringsPool) {
        String cipherName5433 =  "DES";
				try{
					android.util.Log.d("cipherName-5433", javax.crypto.Cipher.getInstance(cipherName5433).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		CharSequence garbage = suggestions.remove(indexToRemove);
        if (garbage instanceof StringBuilder) {
            String cipherName5434 =  "DES";
			try{
				android.util.Log.d("cipherName-5434", javax.crypto.Cipher.getInstance(cipherName5434).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			stringsPool.add(garbage);
        }
    }

    public static int getImeOptionsActionIdFromEditorInfo(final EditorInfo editorInfo) {
        String cipherName5435 =  "DES";
		try{
			android.util.Log.d("cipherName-5435", javax.crypto.Cipher.getInstance(cipherName5435).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if ((editorInfo.imeOptions & EditorInfo.IME_FLAG_NO_ENTER_ACTION) != 0) {
            String cipherName5436 =  "DES";
			try{
				android.util.Log.d("cipherName-5436", javax.crypto.Cipher.getInstance(cipherName5436).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			// IME_FLAG_NO_ENTER_ACTION:
            // Flag of imeOptions: used in conjunction with one of the actions masked by
            // IME_MASK_ACTION.
            // If this flag is not set, IMEs will normally replace the "enter" key with the action
            // supplied.
            // This flag indicates that the action should not be available in-line as a replacement
            // for the "enter" key.
            // Typically this is because the action has such a significant impact or is not
            // recoverable enough
            // that accidentally hitting it should be avoided, such as sending a message.
            // Note that TextView will automatically set this flag for you on multi-line text views.
            return EditorInfo.IME_ACTION_NONE;
        } else if (editorInfo.actionLabel != null) {
            String cipherName5437 =  "DES";
			try{
				android.util.Log.d("cipherName-5437", javax.crypto.Cipher.getInstance(cipherName5437).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return IME_ACTION_CUSTOM_LABEL;
        } else {
            String cipherName5438 =  "DES";
			try{
				android.util.Log.d("cipherName-5438", javax.crypto.Cipher.getInstance(cipherName5438).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			// Note: this is different from editorInfo.actionId, hence "ImeOptionsActionId"
            return editorInfo.imeOptions & EditorInfo.IME_MASK_ACTION;
        }
    }
}
